Revert "Move boiler plate into app.cc"
This reverts commit 0cef37d5b3a0e0d47570926bcf5625330a0c403a.
